Move utility method
